Supply valid Cloudinary upload URL and params
Unfortunately Cloudinary's API requires that the file be uploaded as a
parameter named `file`, so the existing ActiveStorage JS doesn't work
at all. :(

However, since Cloudinary doesn't expect any headers at all, we're free
to use the headers object to hold all of the params that we need to send
along with the file itself. After these changes, we can tweak the JS to
do what we need.